Add content pagination and Fix Active Model Serializers
This commit is necessary because:
In cases where the user has a lot of bucketlists, they should have an option to quickly navigate through the lists. Besides, sending everything all at once is not an entirely nice option.
Active model serialization is not wraping json objects. Render is still using the default rails to_json method

This addresses the need by:
* Creating and implementing a pagination class that handles delivers requested resources in pages
* Pagination class includes resource/page information as meta content in the request sent
* Change the version of Active model serializer used